% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{Build_post_summary}
\alias{Build_post_summary}
\alias{Build_post_summary_TCGA}
\title{Build count matrix for input samples}
\usage{
Build_post_summary_TCGA(ccf_folder, output = NA, minsample = 30)
}
\arguments{
\item{ccf_folder}{ccf files folder}

\item{output}{output path}
}
\value{
a data frame containing the summary for all samples in the ccf files
}
\description{
Build count matrix for input samples
}
