% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{ccfMatBuild}
\alias{ccfMatBuild}
\title{Build CCF Matrix for samples}
\usage{
ccfMatBuild(
  samplenamelist,
  ccf_upper = 1,
  ccf_folder,
  binwidth = 0.01,
  random = FALSE
)
}
\arguments{
\item{ccf_upper}{ccf upper bound}

\item{ccf_folder}{ccf files folder path}

\item{binwidth}{bin width}

\item{random}{output randomized ccf matrix}

\item{samplelist}{samplename list, eg., TCGA-05-1425}
}
\value{
ccfBandCountsMat,ccfBandCountsMat.random,ccfBandFractionMat and ccfBandFractionMat.random
}
\description{
Build CCF Matrix for samples
}
