% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{ccfMatrixBuild}
\alias{ccfMatrixBuild}
\alias{ccfMatBuild_output}
\title{Build count matrix for input samples per cancer type}
\usage{
ccfMatBuild_output(
  ccf_folder,
  output = NA,
  ccf_upper = 1,
  RankEstimateType = "fraction",
  minsample = 30
)
}
\arguments{
\item{ccf_folder}{ccf files folder}

\item{output}{output folder path}

\item{ccf_upper}{CCF upper bound}

\item{RankEstimateType}{output rank estimate matrix format}

\item{minsample}{minimum samples required for each cancer type}
}
\value{
CCF matrix for each cancer type
}
\description{
Build count matrix for input samples per cancer type
}
