% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{chooseNumOfConsensusSigs}
\alias{chooseNumOfConsensusSigs}
\title{chooseNumOfConsensusSigs}
\usage{
chooseNumOfConsensusSigs(combine_sigs, min = 2, max = 10)
}
\arguments{
\item{min}{Minimum clustering number}

\item{max}{Maximum clustering number}

\item{combine_sig}{Combined type-wise signatures across cancer types}
}
\value{
A data frame with suggested clustering numbers for different methods and distances
}
\description{
Choose the Number of Consensus Signatures using Hierarchical Clustering Number Estimate
}
\details{
This function utilizes hierarchical clustering and various clustering indices to estimate
the optimal number of consensus signatures for a given set of type-wise signatures.
}
