% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{cor.table}
\alias{cor.table}
\title{Correlation calculation function}
\usage{
cor.table(df, var_idx, sig_idx, delete_0 = TRUE, min_n = 30)
}
\arguments{
\item{df}{exposure file merged with measures of interest}

\item{va}{Column indexs of variables A in df}

\item{vb}{Column indexs of variables B in df}
}
\value{
A table including correlation r,p,n between variables A and variables B
}
\description{
Correlation calculation function
}
