% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{cor_facet}
\alias{cor_facet}
\title{Correlation calculation function by group and plot association heatmap - 20200915}
\usage{
cor_facet(
  df,
  var_idx,
  sig_idx,
  facet,
  heatmap = FALSE,
  title = "",
  empty_row_delete = FALSE,
  flip = FALSE,
  keep_all = TRUE,
  col_low = "#4a7b94",
  col_high = "#bb5a39",
  scatter = FALSE,
  scatter_var = NA,
  label.x = NA,
  label.y = NA,
  bin.width.x = NA,
  bin.width.y = NA
)
}
\arguments{
\item{df}{exposure file merged with measures of interest}

\item{facet}{name of group variable}

\item{heatmap}{specify whether plot association heatmap}

\item{title}{Title for association heatmap}

\item{empty_row_delete}{whether to delete rows without any significant correlation value}

\item{flip}{whether to flip facet and x}

\item{keep_all}{whether to show all facet}

\item{col_low}{color for correlation r=-1}

\item{col_high}{color for correlation r=1}

\item{va}{Column indexs of variables A in df}

\item{vb}{Column indexs of variables B in df}
}
\value{
A table including correlation r,p,n between variables A and variables B during group variable and association heatmap (optional)
}
\description{
Correlation calculation function by group and plot association heatmap - 20200915
}
