% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{EvoSig_estimate}
\alias{EvoSig_estimate}
\title{Estimate signautrue exposure using consensus signatures of evolutionary dynamics for TCGA data}
\usage{
EvoSig_estimate(ccfMat_by_samples, consensus_sig)
}
\arguments{
\item{ccfMat_by_samples}{CCF matrix by samples}

\item{consensus_sig}{Consensus signatures of evolutionary dynamics}
}
\value{
Exposure matrix
}
\description{
This function estimates exposure using consensus signatures for TCGA data.
}
