% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{EvoSig_generate}
\alias{EvoSig_generate}
\title{EvoSig_generate: Generate Evolutionary Signatures}
\usage{
EvoSig_generate(ccfMat, rank, nrun)
}
\arguments{
\item{ccfMat}{A matrix containing ccf distribution data}

\item{nrun}{Run times}

\item{num_sigs}{The number of signatures suggested by rank estimate analysis}
}
\value{
A list containing results, signatures, and exposures
}
\description{
This function performs NMF for a given cancer type and generates evolutionary signatures.
}
