% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{file_format}
\alias{file_format}
\title{Unify format of data frame}
\usage{
file_format(filename = filename, samplenamecol)
}
\arguments{
\item{samplenamecol}{column index of samplename}

\item{df}{data frame}
}
\description{
Unify format of data frame
}
