% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{hc_combine_sig Combine signatures and estimate number of clusters}
\alias{hc_combine_sig Combine signatures and estimate number of clusters}
\alias{hc_combine_sig}
\title{Combine signatures and estimate number of clusters}
\usage{
hc_combine_sig(
  combine_sigs,
  output_folder,
  min.nc = 3,
  max.nc = 10,
  method = "ward.D2",
  index = "hubert",
  distance = "euclidean"
)
}
\arguments{
\item{output_folder}{output folder path}

\item{min.nc}{minimum cluster}

\item{max.nc}{maximum cluster}

\item{distance}{distance function}

\item{combined_sigs}{combined signatures for all cancer types}
}
\value{
list(combine_sig,cluster)
}
\description{
Combine signatures and estimate number of clusters
}
