% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{scatter_facet}
\alias{scatter_facet}
\title{Plot scatterplot for specific measure with evo sig and subtype - 20200915}
\usage{
scatter_facet(
  var,
  df,
  vb,
  var_idx,
  facet_name = "cancertype",
  cor_table,
  label.x = NA,
  label.y = NA,
  bin.width.x = NA,
  bin.width.y = NA,
  x_lab = "",
  legend.position = "right",
  highlight_color = "#1A9993FF",
  zero_delete = TRUE,
  subtype_panel = FALSE,
  varwidth = FALSE,
  log2_scale = FALSE,
  shape_subtype = FALSE,
  compare_subtype = FALSE,
  side_boxplot = TRUE
)
}
\arguments{
\item{df}{exposure file merged with measures of interest}

\item{vb}{Column indexs of all measures in association heatmap in df}

\item{sig}{Column indexs of evo signature}

\item{facet}{name of facet variable consistent with association heatmap}
}
\value{
Scatterplots
}
\description{
Plot scatterplot for specific measure with evo sig and subtype - 20200915
}
