% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{sig_plot}
\alias{sig_plot}
\alias{ccf_dist}
\title{Plot signature matrix}
\usage{
ccf_dist(
  ccf_rows,
  density_plot = FALSE,
  tag = NULL,
  col = NA,
  theme = "grey",
  fraction = FALSE,
  strip_title = "Signature ",
  title = "Consensus Signature of evolutionary dynamics"
)
}
\arguments{
\item{sig}{signature matrix}
}
\value{
mydata mutation data frame
}
\description{
Plot signature matrix
}
