% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{strip_color}
\alias{strip_color}
\title{Customize top strip color}
\usage{
strip_color(p, col1 = NULL, col2 = NULL, draw = FALSE, direction = "top")
}
\arguments{
\item{p}{plot}

\item{col1}{customized color}

\item{col2}{customized color}

\item{draw}{whether to display in plots panal}

\item{direction}{strip location}
}
\description{
Customize top strip color
}
