% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{combine_sig_nmf}
\alias{combine_sig_nmf}
\title{Combine signatures from different cancer type together}
\usage{
combine_sig_nmf(sig_folder, output_folder = NA, cancertype)
}
\arguments{
\item{sig_folder}{signatures folder}

\item{output_folder}{output_folder}

\item{cancertype}{cancer type list}
}
\value{
matrix combining all signature matrix
}
\description{
Combine signatures from different cancer type together
}
