% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_function.R
\name{strip_color}
\alias{strip_color}
\title{Customize top strip color}
\usage{
strip_color(
  p,
  col1 = signature_col,
  col2 = NULL,
  draw = FALSE,
  direction = "top"
)
}
\arguments{
\item{p}{plot}

\item{draw}{whether to display in plots panal}

\item{direction}{strip location}

\item{col}{customized color}
}
\description{
Customize top strip color
}
